minor refactor - start cc api at v1, not v3
